+++ src/services/publication/information.service.ts
@@ -1,5 +1,7 @@
 import { Information } from "../../domain/models/publication/information.model";
 import { InformationRepository } from "../../domain/repositories/publication/information.repository";
+import { AdministratorService } from "../user/administrator.service";
+import { EditorService } from "../user/editor.service";
 import { ContentService } from "./content.service";
 
 export class InformationService {
@@ -7,6 +9,8 @@ export class InformationService {
     constructor(
         private informationRepository: InformationRepository,
         private contentService: ContentService,
+        private administratorService : AdministratorService,
+        private editorService: EditorService,
     ){};
 
     getAllInformation(): Information[]{
@@ -29,11 +33,23 @@ export class InformationService {
         return information;
     };
 
-    deleteInformation(informationId: number): void {
-        let Information = this.informationRepository.getInformationById(informationId);
-        let InformationContentId = Information.getContent();
-        this.contentService.deleteContent(InformationContentId.getId());
-        this.informationRepository.deleteInformation(informationId);
+    deleteInformation(requestingUserId: number, informationId: number): void | Error {
+        if (this.editorService.isEditor(requestingUserId) || this.administratorService.isAdmin(requestingUserId)){
+            let Information = this.informationRepository.getInformationById(informationId);
+            let InformationContentId = Information.getContent();
+            this.contentService.deleteContent(InformationContentId.getId());
+            this.informationRepository.deleteInformation(informationId);
+        } else {
+            throw new Error('Unauthorized');
+        };
+    };
+
+    changeStatus(requestingUserId: number, informationId: number, newStatus: boolean): void | Error {
+        if (this.administratorService.isAdmin(requestingUserId) || this.editorService.isEditor(requestingUserId)){
+            this.getInformationById(informationId).setStatus(newStatus)
+        } else {
+            throw new Error ('Unauthorized')
+        };
     };
 
 };
