+++ src/services/user/user.service.ts
@@ -1,13 +1,16 @@
 import { Injectable } from '@nestjs/common';
 import { UserRepository } from '../../domain/repositories/user/user.repository';
 import { User } from '../../domain/models/user/user.model';
+import { AdministratorService } from './administrator.service';
 
 
 @Injectable()
 export class UserService {
 
-    constructor (private userRepository: UserRepository
-    ) {};
+    constructor (
+        private userRepository: UserRepository,
+        private administratorService: AdministratorService
+    ){};
 
 
     getAllUsers(): User[] {
@@ -18,19 +21,28 @@ export class UserService {
         return this.userRepository.getUserById(userId);
     };
 
-    createUser(user: User): User {
-        this.userRepository.createUser(user);
-        return user
+    createUser(requestingUserId: number, user: User): User | Error {
+        if (this.administratorService.isAdmin(requestingUserId)) {
+            this.userRepository.createUser(user);
+            return user
+        };
+        throw new Error('Unauthorized');
     };
 
-    editUser(user: User): User {
-        this.userRepository.editUser(user);
-        return user
+    editUser(requestingUserId: number, user: User): User | Error {
+        if (this.administratorService.isAdmin(requestingUserId)) {
+            this.userRepository.editUser(user);
+            return user
+        };
+        throw new Error('Unauthorized');
     };
 
-    deleteUser(userId: number): void {
-        this.userRepository.deleteUser(userId);
+    deleteUser(requestingUserId: number, userId: number): void | Error {
+        if (this.administratorService.isAdmin(requestingUserId)) {
+            this.userRepository.deleteUser(userId);
+        } else {
+            throw new Error('Unauthorized');
+        };
     };
-
 
 };
